internal/monitoring: coarser grained request monitoring
As things were, we were monitoring requests with fairly detailed URL path
information in and this was causing a large number of summary histograms
to be created and hence taking a large amount of memory.

This changes the monitoring to use a coarser grained mechanism
that provides less information but does not allow clients to create
an arbitrary number of histograms.

We also make the Reset/ObserveMetric pairing more obvious,
fix what seems to be a bug where the v4 request monitor wasn't
reset, rename ObserveMetric to Done and add method information
to the request information.

The information recorded for a request now includes the
method, the root of the handler tree responsible for
serving the request, and some handler-specific information
about the request path. All metadata requests now appear
as "meta".