internal/blobstore: implement RemoveUpload
We have to be careful that we (probably RemoveExpiredUploads)
can't remove the parts for an upload just as someone has added the
upload to the store. Unfortunately this case is hard to test, because
it needs the FinishUpload to happen between RemoveUpload retrieving
the document and trying to remove it.