internal/charmstore: calculate next revision number differently
Currently we calculate the next revision of a charm
or bundle by taking the maximum revision of all
the existing matching entities. This becomes a
problem if you want to delete entities because
there's no longer a way to find out what revision
number to assign next.

We change the logic to keep a separate revision
number collection instead so even when a charm
is deleted, we'll still be able to avoid duplicating
old revisions.

Still to come: the migration code.