internal/blobstore: only swift needs to know about container
The concept of "container" was exposed in the ObjectStore
interface, but it's not inherent to an object store - of
the two current implementations, only one needs
to know about containers currently.

This also fixes the MongoDB blobstore implementation
so that it continues to use the same paths as the
currently deployed version, so deploying without
Swift won't cause all the blobs to "disappear".