internal/mongodoc: implement storage scheme for hashes
With the new blob storage scheme, we want to store
the hash of all parts within the mongodoc entities
(because we're not planning to keep refcounts separately)
so the multipart index needs to store the hashes too.

To save space, we store the hashes as a contiguous
byte array (this saves about 10 bytes per index entry)
but we use a Go representation that is more convenient.