all: use hash-addressed blobs
This changes the blobstore API so that all blobs are hash-addressed.
Because the juju/blobstore implementation is not hash-addressed,
we don't use it any more, but instead migrate its names and hashes
to the new blobref collection.

Because blobs are hash-addressed, we add another field to Entity
which holds the hash of the extra data appended to the archive
zip for backward compatibility.

Blobs are no longer reference counted, but will be garbage collected.
Garbage collection is as yet unimplemented and will be done in the
next PR, so tests that rely on blob deletion have been commented
out temporarily until that lands.