internal/blobstore: implement garbage collection
After some experimentation, it seems that holding
all the hashes in memory is a reasonable approach.
Currently there are ~45000 blobs in the store, giving
a projected memory usage of ~2.5MB for a GC.
This implementation should scale reasonably up
to 1 million blobs (20x current usage) and to 10 million
with minor changes. Beyond that, we may need to
use a different algorithm, but that should be OK
to do without changing the database.