internal/{charmstore,v5}: fix list results
When some list results had different ACLs from others,
we filtered out all but the most recent revision, so
we could fail to show any results even though some
results could be shown.

We fix that by changing ListQuery so that it returns
all matching charms that are in different channels;
then we filter those results after retrieving from MongoDB.

This changes the semantics of sorting by owner because
sorting is now done based on the returned results which
don't necessarily contain the actual owner, so we sort by
the User field in the returned URL, which means that
promulgated charms will always be returned first in that case.