internal/monitoring: Fix endpoint metrics.
Here we fix the endpoint metrics so that that they are nicely grouped rather than outputting the entity names and causing time series pollution.
(e.g. /wordpress/meta/common-info, /mysql/meta/common-info etc are now simply tracked as /:id/meta/common-info)
Kind (which was very buggy and non-deterministic) is replaced with Endpoint which holds this above info. The rough metrics on v4 are removed too.