reorder stats key
This means that we can summarize over all series
for a charm in a single request.

We also change the Counters logic to add a fast path for
non-list stats queries.

We also avoid adding ArchiveUploadFailed entries to the stats
as they're the most common entry and we don't need the noise.

Also change the stats granularity to one hour so we get
less records overall.