cmd/migratestats: new command
This migrates the stats collection to the new format.
It uses per-hour granularity, removes categories we
don't want to see any more, and reorders the keys.

This is designed to be run offline while the charm store
is down, but to make sure, we add a couple of migration
steps that will fail until this migration has taken place.